requirements.txt
environs==11.0.0
marshmallow==3.21.2
packaging==24.0
py4j~=0.10.9.7
pyspark==3.5.1
python-dotenv~=1.0.1
mypy~=1.10.0
graphframes~=0.6
pyarrow~=16.1.0
pandas~=2.2.2
numpy==1.26.4
coverage==7.2.7
pytest~=8.3.2
boto3~=1.34.144
botocore~=1.34.144
